#73: final cleanup good. templates
