+++ edu_storybook/admin.py
@@ -20,20 +20,45 @@ admin = Blueprint('admin', __name__)
 
 @admin.route("/admin/")
 def gen_admin_index():
-    return "e"
+    the_index_page = TEMPLATES["_base"].substitute(
+        title = "Admin Homepage",
+        description = "A motivational storybook that helps students learn.",
+        body = TEMPLATES["admin"]["index"].substitute()
+    )
+    return the_index_page
 
 @admin.route("/admin/book_manager")
 def gen_admin_book_mananger():
-    return "e"
+    book_manager_page = TEMPLATES["_base"].substitute(
+        title = "Admin: Book Manager",
+        description = "A motivational storybook that helps students learn.",
+        body = TEMPLATES["admin"]["book_manager"].substitute()
+    )
+    return book_manager_page
 
 @admin.route("/admin/edit_book")
 def gen_admin_edit_book():
-    return "e"
+    edit_book_page = TEMPLATES["_base"].substitute(
+        title = "Admin: Edit Book",
+        description = "A motivational storybook that helps students learn.",
+        body = TEMPLATES["admin"]["edit_book"].substitute()
+    )
+    return edit_book_page
 
 @admin.route("/admin/upload_book")
 def gen_admin_upload_book():
-    return "e"
+    upload_book_page = TEMPLATES["_base"].substitute(
+        title = "Admin: Upload Book",
+        description = "A motivational storybook that helps students learn.",
+        body = TEMPLATES["admin"]["upload_book"].substitute()
+    )
+    return upload_book_page
 
 @admin.route("/admin/study_manager")
-def gen_admin_edit_book():
-    return "e"
+def gen_admin_study_manager():
+    study_manager_page = TEMPLATES["_base"].substitute(
+        title = "Admin: Study Manager",
+        description = "A motivational storybook that helps students learn.",
+        body = TEMPLATES["admin"]["study_manager"].substitute()
+    )
+    return study_manager_page
