+++ edu_storybook/api/__init__.py
@@ -1,0 +1,4 @@
+"""
+__init__.py
+    Initialize edu_storybook.api module.
+"""
