+++ edu_storybook/api/admin.py
@@ -1,14 +1,8 @@
 """
-api.py
-    Contains all the API endpoints.
+admin.py
+    Routes beginning with /api/admin/
 
 Routes:
-    /api/
-    /api/login
-    /api/logout
-    /api/register
-    /api/password/forgot
-    /api/password/reset
     /api/admin/book/download
     /api/admin/book/upload
     /api/admin/book/grant
@@ -16,9 +10,6 @@ Routes:
     /api/admin/download/user
     /api/admin/download/action
     /api/admin/get/user
-    /api/quiz/submit
-    /api/storyboard/page/<int:book_id_in>/<int:page_number_in>
-    /api/storyboard/action
 """
 
 from flask import request
@@ -37,6 +28,7 @@ import random
 import string
 import datetime
 import bcrypt
+import time
 
 from core.auth import validate_login, issue_auth_token
 from core.bucket import bucket
@@ -48,453 +40,9 @@ from core.sensitive import jwt_key
 from core.remove_watchdog import future_del_temp
 from core.reg_exps import *
 
-api = Blueprint('api', __name__)
+a_admin = Blueprint('a_admin', __name__)
 
-@api.route("/api/")
-def api_index():
-    # revalidate login
-    if 'Authorization' in request.cookies:
-        # check if a user has valid credentials
-        auth = request.cookies.get('Authorization')
-        vl = validate_login(
-            auth,
-            permission=0
-        )
-        if vl != True:
-            return vl
-
-        res = make_response({
-            "status": "ok",
-            "login": "reverified"
-        })
-        issue_auth_token(res, auth)
-        return res
-    return {
-        "status": "ok"
-    }
-
-@api.route("/api/login")
-def login(email: str, password: str):
-    # check that all expected inputs are received
-    try:
-        assert 'email' in request.form
-        assert 'password' in request.form
-    except AssertionError:
-        return {
-            "status": "fail",
-            "fail_no": 1,
-            "message": "Either the email or the password was not provided."
-        }, 400, {"Content-Type": "application/json"}
-
-    # sanitize inputs: make sure they're all alphanumeric, longer than 8 chars
-    if re_email.match(request.form['email']) is None or \
-            re_alphanumeric8.match(request.form['password']) is None:
-        return {
-            "status": "fail",
-            "fail_no": 2,
-            "message": "Either the email or the password failed a sanitize check. The POSTed fields should be alphanumeric, longer than 8 characters."
-        }, 400, {"Content-Type": "application/json"}
-
-    # all good, now query database
-    email = (request.form['email']).lower().strip()
-
-    cursor = connection.cursor()
-    try:
-        cursor.execute(
-            "select * from USER_PROFILE where email='" + email + "'"
-        )
-        label_results_from(cursor)
-    except cx_Oracle.Error as e:
-        return {
-            "status": "fail",
-            "fail_no": 3,
-            "message": "Error when querying database.",
-            "database_message": str(e)
-        }, 400, {"Content-Type": "application/json"}
-
-    result = cursor.fetchone()
-    if result is None:
-        return {
-            "status": "fail",
-            "fail_no": 4,
-            "message": "No email matches what was passed."
-        }, 400, {"Content-Type": "application/json"}
-
-    # print(result)
-    # print(result[8])
-    if not bcrypt.checkpw(request.form['password'].encode('utf8'), result['PASSWORD'].encode('utf8')):
-        return {
-            "status": "fail",
-            "fail_no": 5,
-            "message": "Password is incorrect."
-        }, 400, {"Content-Type": "application/json"}
-
-    user_id = result['USER_ID']
-    session_id = str(uuid.uuid4())  # generate a unique token for a user
-
-    try:
-        conn_lock.acquire()
-        cursor.execute(
-            "update USER_SESSION set session_id='" + session_id +
-            "', active=1 where user_id='" + str(user_id) + "'"
-        )
-        connection.commit()
-    except cx_Oracle.Error as e:
-        return {
-            "status": "fail",
-            "fail_no": 6,
-            "message": "Error when updating database.",
-            "database_message": str(e)
-        }, 400, {"Content-Type": "application/json"}
-    finally:
-        conn_lock.release()
-
-    iat = int(time.time())
-
-    res = make_response({
-        "status": "ok",
-        "message": "Successfully authenticated",
-        "iat": iat
-    })
-    token = jwt.encode({
-        "iat": iat,
-        "session": session_id,
-        "sub": user_id,
-        "permission": result['ADMIN']
-    }, jwt_key, algorithm=config['jwt_alg'])
-    res.set_cookie(
-        "Authorization",
-        "Bearer " + token,
-        max_age=config["login_duration"],
-        # domain=domain_name#, # TODO: uncomment in production
-        # secure=True,
-        # httponly=True
-    )
-
-    try:
-        conn_lock.acquire()
-        cursor.execute(
-            "update USER_PROFILE set LAST_LOGIN=CURRENT_TIMESTAMP where user_id='" +
-            str(user_id) + "'"
-        )
-        connection.commit()
-    except cx_Oracle.Error as e:
-        return {
-            "status": "fail",
-            "fail_no": 7,
-            "message": "Error when updating database.",
-            "database_message": str(e)
-        }, 400, {"Content-Type": "application/json"}
-    finally:
-        conn_lock.release()
-
-    return res
-
-@api.route("/api/logout")
-def logout(auth):
-    # make sure the user is authenticated first
-    auth = request.cookies.get('Authorization')
-    vl = validate_login(
-        auth,
-        permission=0
-    )
-    if vl != True:
-        return vl
-
-    if 'Bearer ' in auth:
-        auth = auth.replace('Bearer ', '', 1)
-    token = jwt.decode(auth, jwt_key, algorithms=config['jwt_alg'])
-
-    cursor = connection.cursor()
-    try:
-        conn_lock.acquire()
-        cursor.execute(
-            "update USER_SESSION set active=0 where user_id='" +
-            token['sub'] + "'"
-        )
-        connection.commit()
-    except cx_Oracle.Error as e:
-        return {
-            "status": "fail",
-            "fail_no": 8,
-            "message": "Error when updating database.",
-            "database_message": str(e)
-        }
-    finally:
-        conn_lock.release()
-
-    res = make_response({
-        "status": "ok"
-    })
-    res.set_cookie('Authorization', '', expires=0)
-    return res
-
-@api.route("/api/register")
-def register(email: str, password: str, first_name: str, last_name: str, school_id: int):
-    # check that all expected inputs are not empty
-    try:
-        assert len('email') > 0
-        assert len('password') > 0
-        assert len('first_name') > 0
-        assert len('last_name') > 0
-        assert len('school_id') > 0
-    except AssertionError:
-        return {
-            "status": "fail",
-            "fail_no": 1,
-            "message": "A field was not provided"
-        }
-
-    # sanitize inputs: make sure they're all alphanumeric, longer than 8 chars
-    if re_email.match(request.form['email']) is None or \
-            re_alphanumeric8.match(request.form['password']) is None or \
-            re_alphanumeric2.match(request.form['first_name']) is None or \
-            re_alphanumeric2.match(request.form['last_name']) is None:
-        return {
-            "status": "fail",
-            "fail_no": 2,
-            "message": "Some field failed a sanitize check. The POSTed fields should be alphanumeric, longer than 8 characters."
-        }
-
-    # all good, now query database
-    email = (email).lower().strip()
-    first_name = (first_name).strip()
-    last_name = (last_name).strip()
-    school_id = (school_id).lower().strip()
-
-    cursor = connection.cursor()
-    try:
-        cursor.execute(
-            "select * from USER_PROFILE where email='" + email + "'"
-        )
-    except cx_Oracle.Error as e:
-        return {
-            "status": "fail",
-            "fail_no": 3,
-            "message": "Error when querying database.",
-            "database_message": str(e)
-        }
-
-    result = cursor.fetchone()
-    if result is not None:
-        return {
-            "status": "fail",
-            "fail_no": 4,
-            "message": "Email is Already Registered."
-        }
-
-    hashed = bcrypt.hashpw(
-        request.form['password'].encode('utf8'), bcrypt.gensalt())
-
-    try:
-        conn_lock.acquire()
-        cursor.execute(
-            "INSERT into USER_PROFILE (email, first_name, last_name, admin, school_id, study_id, password) VALUES ('"
-            + email + "', '"
-            + first_name + "', '"
-            + last_name + "', "
-            + "0 , "
-            + school_id + ", "
-            + 'null' + ", '"
-            + hashed.decode('utf8')
-            + "')"
-        )
-        connection.commit()
-    except cx_Oracle.Error as e:
-        return {
-            "status": "fail",
-            "fail_no": 5,
-            "message": "Error when querying database.",
-            "database_message": str(e)
-        }
-    finally:
-        conn_lock.release()
-
-        send_email(first_name + last_name, email, 'Edu Storybooks', 'edustorybooks@gmail.com',
-                   'Welcome to Edu Storybooks', 'Dear ' + first_name + ' ' + last_name + ',' +
-                   '\n\nThanks for registering an account with Edu Storybooks! :)')
-    return {
-        "status": "ok"
-    }
-
-
-# input email & check if email exists
-@api.route("/api/password/forgot", methods=['POST'])
-def password_forgot():
-
-    # checks for input
-    try:
-        assert 'email' in request.form
-    except AssertionError:
-        return {
-            "status": "fail",
-            "fail_no": 1,
-            "message": "Email was not provided."
-        }, 400, {"Content-Type": "application/json"}
-
-    # sanitize inputs: alphanumeric, > 8 chars
-    if re_email.match(request.form['email']) is None:
-        return {
-            "status": "fail",
-            "fail_no": 2,
-            "message": "Email failed sanitization check of more than 8 characters &/or alphanumeric."
-        }, 400, {"Content-Type": "application/json"}
-
-    # begin querying database
-    email = (request.form['email']).lower().strip()
-
-    # create random sequence of 512 byte string
-    rand_str = ''.join(random.SystemRandom().choice(
-        string.ascii_uppercase + string.digits) for _ in range(512))
-
-    cursor = connection.cursor()
-    try:
-        cursor.execute(
-            # fix the SQL statement with user_session?
-            "SELECT * FROM USER_PROFILE WHERE EMAIL ='" + email + "'")
-    except cx_Oracle.Error as e:
-        return {
-            "status": "fail",
-            "fail_no": 3,
-            "message": "Error when querying database.",
-            "database_message": str(e)
-        }, 400, {"Content-Type": "application/json"}
-
-    result = cursor.fetchone()
-    if result is None:
-        return {
-            "status": "fail",
-            "fail_no": 4,
-            "message": "No email matches what was passed."
-        }, 400, {"Content-Type": "application/json"}
-
-    user_id = result[9]
-    user_name = result[1] + ' ' + result[2]
-
-    now = datetime.datetime.now()
-    req_date = (now.strftime("%Y/%m/%d"))
-
-    try:  # it does not insert with Oracle db? (but works in SQLDeveloper)
-        conn_lock.acquire()
-        cursor.execute(
-            "INSERT INTO PASSWORD_RESET(USER_ID, RESET_KEY, REQUEST_DATE) VALUES('" + user_id + "','" + rand_str + "', TO_DATE('" + req_date + "', 'yyyy/mm/dd'))")
-        connection.commit()
-    except cx_Oracle.Error as e:
-        return {
-            "status": "fail",
-            "fail_no": 5,
-            "message": "Error when querying database.",
-            "database_message": str(e)
-        }, 400, {"Content-Type": "application/json"}
-    finally:
-        conn_lock.release()
-
-    key = 'edustorybook.com/Password/Reset#key=' + rand_str
-
-    send_email(user_name, email, 'Edu Storybooks',
-               'edustorybooks@gmail.com', 'Password Reset Request', key)
-
-    return {
-        "status": "ok"
-    }
-
-
-# new password updates old password in USER_PROFILE & deletes the inserted row in PASSWORD_RESET
-# check if both password fields match
-@api.route("/api/password/reset", methods=['POST'])
-def password_reset():
-   # check expected input
-    try:
-        assert 'new_pass' in request.form
-        assert 'confirm_pass' in request.form
-        assert 'reset_key' in request.form
-    except AssertionError:
-        return {
-            "status": "fail",
-            "fail_no": 1,
-            "message": "Either password was not provided."
-        }, 400, {"Content-Type": "application/json"}
-
-    # sanitize inputs: make sure they're all alphanumeric, longer than 8 chars
-    if re_alphanumeric8.match(request.form['new_pass']) is None or \
-            re_alphanumeric8.match(request.form['confirm_pass']) is None:
-        return {
-            "status": "fail",
-            "fail_no": 2,
-            "message": "Either one or both passwords failed sanitization check of more than 8 characters &/or alphanumeric."
-        }, 400, {"Content-Type": "application/json"}
-
-    # check if both passwords match
-    if (request.form['new_pass'] != request.form['confirm_pass']):
-        return {
-            "status": "fail",
-            "fail_no": 3,
-            "message": "Both passwords do not match."
-        }, 400, {"Content-Type": "application/json"}
-
-    hashed = bcrypt.hashpw(
-        request.form['confirm_pass'].encode('utf8'), bcrypt.gensalt())
-
-    reset_key = (request.form['reset_key'])
-
-    # connect to database
-    cursor = connection.cursor()
-    try:
-        cursor.execute(
-            "SELECT USER_ID FROM PASSWORD_RESET WHERE RESET_KEY ='" + reset_key + "'")
-
-    except cx_Oracle.Error as e:
-        return {
-            "status": "fail",
-            "fail_no": 4,
-            "message": "Error when querying database.",
-            "database_message": str(e)
-        }, 400, {"Content-Type": "application/json"}
-
-    result = cursor.fetchone()
-    if result is None:
-        return {
-            "status": "fail",
-            "fail_no": 5,
-            "message": "No reset_key matches what was passed."
-        }, 400, {"Content-Type": "application/json"}
-
-    try:
-        conn_lock.acquire()
-        cursor.execute("UPDATE USER_PROFILE set PASSWORD ='" +
-                       hashed.decode('utf8') + "' WHERE user_id ='" + result[0] + "'")
-        connection.commit()
-    except cx_Oracle.Error as e:
-        return {
-            "status": "fail",
-            "fail_no": 6,
-            "message": "Error when querying database.",
-            "database_message": str(e)
-        }, 400, {"Content-Type": "application/json"}
-    finally:
-        conn_lock.release()
-
-    try:
-        conn_lock.acquire()
-        cursor.execute(
-            "DELETE FROM PASSWORD_RESET WHERE RESET_KEY ='" + reset_key + "'")
-        connection.commit()
-    except cx_Oracle.Error as e:
-        return {
-            "status": "fail",
-            "fail_no": 7,
-            "message": "Error when querying database.",
-            "database_message": str(e)
-        }, 400, {"Content-Type": "application/json"}
-    finally:
-        conn_lock.release()
-
-    return {
-        "status": "ok"
-    }
-
-
-@api.route("/api/admin/book/download", methods=['POST'])
+@a_admin.route("/api/admin/book/download", methods=['POST'])
 def admin_download_book():
 
     # validate that user has admin rights to download books
@@ -521,7 +69,7 @@ def admin_download_book():
         }, 400, {"Content-Type": "application/json"}
 
 
-@api.route("/api/admin/book/upload", methods=['POST'])
+@a_admin.route("/api/admin/book/upload", methods=['POST'])
 def admin_book_upload():
 
     # validate that user has admin rights to upload books
@@ -653,7 +201,7 @@ def admin_book_upload():
     }, 400, {"Content-Type": "application/json"}
 
 
-@api.route("/api/admin/book/grant", methods=['POST'])
+@a_admin.route("/api/admin/book/grant", methods=['POST'])
 def admin_add_book_to_study():
 
     # validate that user can access data
@@ -699,7 +247,7 @@ def admin_add_book_to_study():
         conn_lock.release()
 
 
-@api.route("/api/admin/page", methods=['POST', 'GET', 'PUT', 'DELETE'])
+@a_admin.route("/api/admin/page", methods=['POST', 'GET', 'PUT', 'DELETE'])
 def admin_page_handler():
     """
     This endpoint handles quiz questions and answers
@@ -930,7 +478,7 @@ def admin_page_handler():
         return "Invalid Operation"
 
 
-@api.route("/api/admin/download/user", methods=['POST'])
+@a_admin.route("/api/admin/download/user", methods=['POST'])
 def admin_download_user_data():
     """
     Exports user profile data to a csv file
@@ -1024,7 +572,7 @@ def admin_download_user_data():
         }
 
 
-@api.route("/api/admin/download/action", methods=['POST'])
+@a_admin.route("/api/admin/download/action", methods=['POST'])
 def admin_download_action_data():
     """
     Exports user action data to a csv file
@@ -1119,7 +667,7 @@ def admin_download_action_data():
 
 
 # take in input param ofset that will be the limit of 50 ofset of 50 and then be happy.
-@api.route("/api/admin/get/user", methods=['GET'])
+@a_admin.route("/api/admin/get/user", methods=['GET'])
 def admin_get_users():
     """
     Exports user data to a json
@@ -1189,213 +737,4 @@ def admin_get_users():
     return {
         "status": "ok",
         "users": users
-    }
-
-@api.route("/api/quiz/submit", methods=['POST'])
-def quiz_submit_answer():
-    try:
-        assert 'answer_id' in request.form
-        assert 'question_id' in request.form
-    except AssertionError:
-        return {
-            "status": "fail",
-            "fail_no": 4,
-            "message": "Either the answer_id or the question_id was not provided."
-        }, 400, {"Content-Type": "application/json"}
-
-    try:
-        answer_id = int(request.form['answer_id'])
-        question_id = int(request.form['question_id'])
-    except ValueError:
-        return {
-            "status": "fail",
-            "fail_no": 5,
-            "message": "Either the answer_id or the question_id contained invalid characters."
-        }, 400, {"Content-Type": "application/json"}
-
-    # make sure the user is authenticated first
-    auth = request.cookies.get('Authorization')
-    vl = validate_login(
-        auth,
-        permission=0
-    )
-    if vl != True:
-        return vl
-
-    if 'Bearer ' in auth:
-        auth = auth.replace('Bearer ', '', 1)
-    token = jwt.decode(auth, jwt_key, algorithms=config['jwt_alg'])
-
-    cursor = connection.cursor()
-    try:
-        conn_lock.acquire()
-        cursor.execute(
-            "insert into user_response (user_id, question_id, answer_id, answered_on) values (" + "'" +
-            token['sub'] + "', " + str(question_id) +
-            ", " + str(answer_id) + ", current_timestamp)"
-        )
-        connection.commit()
-    except cx_Oracle.Error as e:
-        return {
-            "status": "fail",
-            "fail_no": 6,
-            "message": "Error when updating database.",
-            "database_message": str(e)
-        }, 400, {"Content-Type": "application/json"}
-    finally:
-        conn_lock.release()
-
-    return {"status": "ok"}
-
-
-@api.route("/api/storyboard/page/<int:book_id_in>/<int:page_number_in>", methods=['GET'])
-def storyboard_get_page(book_id_in, page_number_in):
-    # make sure user is authenticated
-    auth = request.cookies.get('Authorization')
-    vl = validate_login(
-        auth,
-        permission=0
-    )
-    if vl != True:
-        return vl
-
-    if 'Bearer' in auth:
-        auth = auth.replace('Bearer ', '', 1)
-
-    token = jwt.decode(auth, jwt_key, algorithms=config['jwt_alg'])
-
-    # sanitize inputs: make sure book_id, page_number are ints
-    try:
-        book_id = int( book_id_in)
-        page_number = int(page_number_in)
-    except ValueError:
-        return {
-            "status": "fail",
-            "fail_no": 2,
-            "message": "The book_id or page_number failed a sanitize check. The POSTed fields should be an integer."
-        }, 400, {"Content-Type": "application/json"}
-
-    # goes into database and gets the bucket folder.
-    # goes into bucket and then says I want this image from this folder.
-    cursor = connection.cursor()
-
-    try:
-        # get folder that holds that book's images
-        cursor.execute(
-            "SELECT folder FROM BOOK where book_id =" + str(book_id))
-    except cx_Oracle.Error as e:
-        return {"status": "fail",
-                "fail_no": 3,
-                "message": "Error when updating database action",
-                "database_message": str(e)
-                }, 400, {"Content-Type": "application/json"}
-
-    # check if this is in write format ; then we have to fix it and ammend it with page number
-    fileInput = cursor.fetchone()[0]
-    fileInput = fileInput + '/' + fileInput + '_' + str(page_number) + '.png'
-
-    try:
-        return send_file(bucket.download_bucket_file(fileInput))
-    except:
-        return {
-            "status": "fail",
-            "fail_no": 4,
-            "message": "could not get image"
-        }, 400, {"Content-Type": "application/json"}
-
-
-@api.route("/api/storyboard/action", methods=['POST'])
-def storyboard_save_user_action():
-    '''
-    Code the storyboard_save_user_action() function in the same style as logout()
-    Screenshot a successful POST request to the /storyboard/action endpoint.
-    Add to the backend/API.md the relevant documentation for the /storyboard/action endpoint
-    in the same style as the login/ endpoint documentation.
-    '''
-    # make sure user is authenticated
-    auth = request.cookies.get('Authorization')
-    vl = validate_login(
-        auth,
-        permission=0
-    )
-    if vl != True:
-        return vl
-
-    if 'Bearer' in auth:
-        auth = auth.replace('Bearer ', '', 1)
-
-    token = jwt.decode(auth, jwt_key, algorithms=config['jwt_alg'])
-
-    # check that all expected inputs are received
-    try:
-        assert 'book_id' in request.form
-        assert 'detail_description' in request.form
-        assert 'action_key_id' in request.form
-        assert 'action_start' in request.form
-        assert 'action_stop' in request.form
-    except AssertionError:
-        return {
-            "status": "fail",
-            "fail_no": 1,
-            "message": "Either the book_id, detail_description, or action_id was not provided."
-        }, 400, {"Content-Type": "application/json"}
-
-    # sanitize inputs: make sure book_id, action_key_id are ints
-    try:
-        book_id = int(request.form["book_id"])
-        action_key_id = int(request.form["action_key_id"])
-    except ValueError:
-        return {
-            "status": "fail",
-            "fail_no": 2,
-            "message": "The book_id or action_key_id failed a sanitize check. The POSTed fields should be an integer for book_id or action_id."
-        }, 400, {"Content-Type": "application/json"}
-
-    # sanitize inputs: make sure action_start and action_stop are in correct format
-    if re_timestamp.match(request.form["action_start"]) is None or \
-            re_timestamp.match(request.form["action_stop"]) is None or \
-            re_alphanumeric.match(request.form["detail_description"]) is None:
-        return {
-            "status": "fail",
-            "fail_no": 3,
-            "message": "Either the action_start, action_stop, or detail_description failed a sanitize check. The POSTed fields should be in date format YYYY-MM-DD HH:MM:SS. detail_description should be alphanumeric only."
-        }, 400, {"Content-Type": "application/json"}
-
-    cursor = connection.cursor()
-    try:
-        cursor.execute(
-            "DECLARE " +
-            "USER_ID_IN VARCHAR2(36);" +
-            "ACTION_START_IN DATE;" +
-            "ACTION_STOP_IN DATE;" +
-            "BOOK_ID_IN NUMBER;" +
-            "DETAIL_DESCRIPTION_IN VARCHAR2(100);" +
-            "ACTION_KEY_ID_IN NUMBER;" +
-            "BEGIN " +
-            "USER_ID_IN := '" + token["sub"]+"'; " +
-            "ACTION_START_IN := TO_DATE('" + request.form["action_start"]+"', 'YYYY-MM-DD HH24:MI:SS'); " +
-            "ACTION_STOP_IN := TO_DATE('" + request.form["action_stop"]+"',  'YYYY-MM-DD HH24:MI:SS'); " +
-            "BOOK_ID_IN := " + request.form["book_id"]+"; " +
-            "DETAIL_DESCRIPTION_IN := '" + request.form["detail_description"] + "'; " +
-            "ACTION_KEY_ID_IN := " + request.form["action_key_id"] + "; " +
-            "CHECK_DETAIL_ID_PROC (" +
-            "USER_ID_IN => USER_ID_IN, " +
-            "ACTION_START_IN => ACTION_START_IN, " +
-            "ACTION_STOP_IN => ACTION_STOP_IN, " +
-            "BOOK_ID_IN => BOOK_ID_IN, " +
-            "DETAIL_DESCRIPTION_IN => DETAIL_DESCRIPTION_IN, " +
-            "ACTION_KEY_ID_IN => ACTION_KEY_ID_IN " +
-            ");" +
-            "END;"
-        )
-        connection.commit()
-    except cx_Oracle.Error as e:
-        return{
-            "status": "fail",
-            "fail_no": 4,
-            "message": "Error when updating database action",
-            "database_message": str(e)
-        }, 400, {"Content-Type": "application/json"}
-    return {
-        "status": "ok"
     }
