+++ edu_storybook/api/quiz.py
@@ -1,0 +1,77 @@
+"""
+quiz.py
+    Routes beginning with /api/quiz/
+
+Routes:
+    /api/quiz/submit
+"""
+
+from flask import request
+from flask import Blueprint
+
+import jwt
+import cx_Oracle
+
+from core.auth import validate_login, issue_auth_token
+from core.config import config
+from core.db import connection, conn_lock
+from core.sensitive import jwt_key
+from core.reg_exps import *
+
+a_quiz = Blueprint('a_quiz', __name__)
+
+@a_quiz.route("/api/quiz/submit", methods=['POST'])
+def quiz_submit_answer():
+    try:
+        assert 'answer_id' in request.form
+        assert 'question_id' in request.form
+    except AssertionError:
+        return {
+            "status": "fail",
+            "fail_no": 4,
+            "message": "Either the answer_id or the question_id was not provided."
+        }, 400, {"Content-Type": "application/json"}
+
+    try:
+        answer_id = int(request.form['answer_id'])
+        question_id = int(request.form['question_id'])
+    except ValueError:
+        return {
+            "status": "fail",
+            "fail_no": 5,
+            "message": "Either the answer_id or the question_id contained invalid characters."
+        }, 400, {"Content-Type": "application/json"}
+
+    # make sure the user is authenticated first
+    auth = request.cookies.get('Authorization')
+    vl = validate_login(
+        auth,
+        permission=0
+    )
+    if vl != True:
+        return vl
+
+    if 'Bearer ' in auth:
+        auth = auth.replace('Bearer ', '', 1)
+    token = jwt.decode(auth, jwt_key, algorithms=config['jwt_alg'])
+
+    cursor = connection.cursor()
+    try:
+        conn_lock.acquire()
+        cursor.execute(
+            "insert into user_response (user_id, question_id, answer_id, answered_on) values (" + "'" +
+            token['sub'] + "', " + str(question_id) +
+            ", " + str(answer_id) + ", current_timestamp)"
+        )
+        connection.commit()
+    except cx_Oracle.Error as e:
+        return {
+            "status": "fail",
+            "fail_no": 6,
+            "message": "Error when updating database.",
+            "database_message": str(e)
+        }, 400, {"Content-Type": "application/json"}
+    finally:
+        conn_lock.release()
+
+    return {"status": "ok"}
