+++ edu_storybook/api/study.py
@@ -1,0 +1,7 @@
+"""
+study.py
+    Handle all study related GET, POST, PUT, DELETE
+
+"""
+
+# ...
