+++ edu_storybook/password.py
@@ -1,4 +1,7 @@
 """
+password.py
+    Generates the webpages for the "/password/" pages.
+
 Routes:
     /password/forgot
     /password/reset
@@ -13,8 +16,18 @@ password = Blueprint('password', __name__)
 
 @password.route("/password/forgot")
 def gen_password_forgot():
-    return "e"
+    password_forgot_page = TEMPLATES["_base"].substitute(
+        title = "Password Forgot",
+        description = "Enter your email to request a password reset",
+        body = TEMPLATES["password"]["forgot"].substitute()
+    )
+    return password_forgot_page
 
 @password.route("/password/reset")
 def gen_password_reset():
-    return "e"
+    password_reset_page = TEMPLATES["_base"].substitute(
+        title = "Password Reset",
+        description = "Reset your password",
+        body = TEMPLATES["password"]["reset"].substitute()
+    )
+    return password_reset_page
