+++ edu_storybook/story_selection.py
@@ -13,6 +13,27 @@ from templates import TEMPLATES
 
 story_selection = Blueprint('story_selection', __name__)
 
-@story_selection.route("/books")
+@story_selection.route('/books')
 def gen_books():
-    return "e"
+    """
+    example:
+    all_books = ""
+    for b in book_query:
+        books += TEMPLATES['story_selection']['book'].substitute(
+            book_title=b['TITLE'],
+            ...
+        )
+    story_selection_page = TEMPLATES["_base"].substitute(
+        title = 'Book Selection',
+        description = 'Select a book to read',
+        body = TEMPLATES['story_selection']['index'].substitute(
+            books=all_books
+        )
+    )
+    """
+    story_selection_page = TEMPLATES["_base"].substitute(
+        title = 'Book Selection',
+        description = 'Select a book to read',
+        body = TEMPLATES['story_selection']['index'].substitute()
+    )
+    return story_selection_page
