+++ edu_storybook/templates/admin/book_manager.html
@@ -1,3 +1,8 @@
+<!--
 ... book_manager.html ...
 
-$books
+TODO: Display all books, and allow the admin to assign books to studies
+suggestion from: @lisapat
+
+~books
+-->
