+++ edu_storybook/templates/admin/edit_book.html
@@ -1,3 +1,6 @@
+<!--
+  /admin/edit_book
+-->
 <div>
 
     <div class="col">
