+++ edu_storybook/templates/admin/index.html
@@ -1,4 +1,9 @@
-... index.html ...
+<!-- ... index.html ...
+
+/admin/
+
+TODO: This is the admin (home)page that will link to all of the other pages
+(book_manager, edit_book, etc.)
 
 
 :^)
@@ -27,3 +32,4 @@
                          &.&.%(%&&&. *&...&(#(&(*.%....(                               
                           &.&**.....&     &(((&.&...*.                                 
                             &.......&        &,@..&   *&                              
+-->
