+++ edu_storybook/templates/admin/study_manager.html
@@ -1,1 +1,8 @@
-... study_manager.html ...
+<!-- ... study_manager.html ...
+''' For moving users and assigning them to studies; additionally, 
+there will be a popup that allows the administrator to create studies 
+(with study_names and study_invite_codes)
+
+suggestion from @lisapat
+
+    -->
