+++ edu_storybook/templates/index.html
@@ -1,3 +1,10 @@
+<style type="text/css">
+  .carousel-control-next,
+  .carousel-control-prev /*, .carousel-indicators */ {
+      filter: invert(0%);
+  }
+</style>
+
 <nav class="navbar navbar-expand-lg navbar-light bg-light">
     <a class="navbar-brand" href="/">EduStorybook</a>
     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
@@ -7,10 +14,10 @@
     <div class="collapse navbar-collapse" id="navbarSupportedContent">
       <ul class="navbar-nav mr-auto">
         <li class="nav-item active">
-          <a href="/Login" class="nav-link">Login</a>
+          <a href="/login" class="nav-link">Login</a>
         </li>
         <li class="nav-item">
-          <a href="/StorySelection" class="nav-link disabled">Story Selection</a>
+          <a href="/books" class="nav-link disabled">Story Selection</a>
         </li>
       </ul>
     </div>
@@ -25,11 +32,11 @@
       </ol>
       <div class="carousel-inner" role="listbox">
         <div class="carousel-item active" data-interval="10000">
-          <img class="d-block w-100 mx-auto" src= {process.env.PUBLIC_URL +'/images/TitlePage.png'}
+          <img class="d-block w-100 mx-auto" src="/static/images/TitlePage.png">
           alt="First Slide"/>
         </div>   
         <div class="carousel-item">
-          <img class="d-block w-100 mx-center" src='/static/Book.jpg'}
+          <img class="d-block w-100 mx-center" src='/static/images/Book.jpg'}
             alt="Second Slide"/>
           <div  id = "Purpose" class="carousel-caption d-none d-md-block mx-auto bg-transparent ">
             <h1> The purpose of this website is to create an environment for learning 
@@ -58,8 +65,8 @@
     </div>
 
     <div className="text-center mt-5">  
-      <a className="btn btn-link" href="/Login">Login</a>
-      <a className="btn btn-link" href="/Register">Register</a>
+      <a className="btn btn-link" href="/login">Login</a>
+      <a className="btn btn-link" href="/register">Register</a>
     </div>
 
 </div>
