+++ edu_storybook/templates/navbar/logged_admin.html
@@ -1,0 +1,1 @@
+...
