+++ edu_storybook/templates/navbar/logged_out.html
@@ -1,0 +1,1 @@
+...
