+++ edu_storybook/templates/navbar/logged_user.html
@@ -1,0 +1,1 @@
+...
