+++ edu_storybook/templates/story_selection/book.html
@@ -1,13 +1,13 @@
-<a href="$book_link">
+<a href="~book_link">
   <div>
     <div>
-      $book_title
+      ~book_title
     </div>
     <div>
-      <img src="$book_image"></img>
+      <img src="~book_image"></img>
     </div>
     <div>
-      $book_description
+      ~book_description
     </div>
   </div>
 </a>
