+++ edu_storybook/templates/story_selection/index.html
@@ -1,9 +1,9 @@
 <div>
     <div>
-      {/*Navigation Bar Here*/}
+      <!-- Navigation Bar Here -->
       <h1>Story Selection</h1>
       <h2>Choose a Story to Read:</h2>
-      <div>{/*Buffer Space*/}</div>
-      $books
+      <div><!-- Buffer Space --></div>
+      ~books
     </div>
 </div>
